Change total size and current size from int to long

+++ app/src/main/java/com/yurii/youtubemusic/screens/youtube/models/Progress.kt
@@ -9,13 +9,13 @@ import java.io.Serializable
  * [currentSize] - Current loaded size in bytes. [currentSizeInMb] returns [currentSize] in megabytes
  * [totalSize] - Total size of downloading item in bytes. [totalSizeInMb] returns [totalSize] in megabytes
  */
-data class Progress(@IntRange(from = 0, to = 100) var progress: Int, var currentSize: Int, var totalSize: Int) : Serializable {
+data class Progress(@IntRange(from = 0, to = 100) var progress: Int, var currentSize: Long, var totalSize: Long) : Serializable {
     val currentSizeInMb: Double
         get() = currentSize.toDouble() / 1000_000
     val totalSizeInMb: Double
         get() = totalSize.toDouble() / 1000_000
 
-    fun update(progress: Int, currentSize: Int, totalSize: Int) {
+    fun update(progress: Int, currentSize: Long, totalSize: Long) {
         this.progress = progress
         this.currentSize = currentSize
         this.totalSize = totalSize
